requirements.txt
python-dotenv==1.0.1
tenacity!=8.4.0
langchain==0.3.9
openai==1.56.0
langchain-openai==0.2.10
neo4j==5.27.0
langchainhub==0.1.21
langchain-neo4j==0.1.1
langchain-community==0.3.9
langchain-experimental==0.3.3
youtube-search==2.1.2
